requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
llama_index
openai


